Audio callback function is done. It isn't choppy anymore because of bad performance or faulty buffer memsetting.
Still, though, audio requries cleaning.
In addition, I removed some things from the SIGINT signal handler and it is now the game developer's responsibility to call TerminateAudio and ResetTerminal. I did this to prevent the game quitting-functions getting called twice by accident (once by the signal handler and another time by the developer) which would break things.

+++ engine/structs.cpp
@@ -28,19 +28,19 @@ void Engine::TimePoint::SetToNow()
 {
 	chornoTimePoint = std::chrono::high_resolution_clock::now();
 }
-int Engine::TimePoint::Seconds()
+long Engine::TimePoint::Seconds()
 {
 	return std::chrono::duration_cast<std::chrono::seconds>(chornoTimePoint - engineStartTP.chornoTimePoint).count();
 }
-int Engine::TimePoint::Milliseconds()
+long Engine::TimePoint::Milliseconds()
 {
 	return std::chrono::duration_cast<std::chrono::milliseconds>(chornoTimePoint - engineStartTP.chornoTimePoint).count();
 }
-int Engine::TimePoint::Microseconds()
+long Engine::TimePoint::Microseconds()
 {
 	return std::chrono::duration_cast<std::chrono::microseconds>(chornoTimePoint - engineStartTP.chornoTimePoint).count();
 }
-int Engine::TimePoint::Nanoseconds()
+long Engine::TimePoint::Nanoseconds()
 {
 	return std::chrono::duration_cast<std::chrono::nanoseconds>(chornoTimePoint - engineStartTP.chornoTimePoint).count();
 }
